Rename RDDRONE-UAVCAN146 board to UCANS32K146
