+++ boards/Kconfig
@@ -1465,14 +1465,14 @@ config ARCH_BOARD_RDDRONE_UAVCAN144
 		This options selects support for NuttX on the NXP RDDRONE-UAVCAN board
 		featuring the S32K144 Cortex-M4F.
 
-config ARCH_BOARD_RDDRONE_UAVCAN146
-	bool "NXP RDDRONE-UAVCAN146"
+config ARCH_BOARD_UCANS32K146
+	bool "NXP UCANS32K146"
 	depends on ARCH_CHIP_S32K146
 	select ARCH_HAVE_LEDS
 	select ARCH_HAVE_BUTTONS
 	select ARCH_HAVE_IRQBUTTONS
 	---help---
-		This options selects support for NuttX on the NXP RDDRONE-UAVCAN board
+		This options selects support for NuttX on the NXP UCANS32K board
 		featuring the S32K146 Cortex-M4F.
 
 config ARCH_BOARD_S32K146EVB
@@ -2475,7 +2475,7 @@ config ARCH_BOARD
 	default "s32k118evb"               if ARCH_BOARD_S32K118EVB
 	default "s32k144evb"               if ARCH_BOARD_S32K144EVB
 	default "rddrone-uavcan144"        if ARCH_BOARD_RDDRONE_UAVCAN144
-	default "rddrone-uavcan146"        if ARCH_BOARD_RDDRONE_UAVCAN146
+	default "ucans32k146"              if ARCH_BOARD_UCANS32K146
 	default "rv32m1-vega"              if ARCH_BOARD_RV32M1_VEGA
 	default "s32k146evb"               if ARCH_BOARD_S32K146EVB
 	default "s32k148evb"               if ARCH_BOARD_S32K148EVB
@@ -2637,8 +2637,8 @@ endif
 if ARCH_BOARD_RDDRONE_UAVCAN144
 source "boards/arm/s32k1xx/rddrone-uavcan144/Kconfig"
 endif
-if ARCH_BOARD_RDDRONE_UAVCAN146
-source "boards/arm/s32k1xx/rddrone-uavcan146/Kconfig"
+if ARCH_BOARD_UCANS32K146
+source "boards/arm/s32k1xx/ucans32k146/Kconfig"
 endif
 if ARCH_BOARD_S32K146EVB
 source "boards/arm/s32k1xx/s32k146evb/Kconfig"
