+++ boards/arm/s32k1xx/ucans32k146/Kconfig
@@ -3,6 +3,6 @@
 # see the file kconfig-language.txt in the NuttX tools repository.
 #
 
-if ARCH_BOARD_RDDRONE_UAVCAN146
+if ARCH_BOARD_UCANS32K146
 
-endif # ARCH_BOARD_RDDRONE_UAVCAN146
+endif # ARCH_BOARD_UCANS32K146
