+++ boards/arm/s32k1xx/ucans32k146/configs/nsh/defconfig
@@ -12,8 +12,8 @@
 # CONFIG_NSH_CMDOPT_HEXDUMP is not set
 # CONFIG_NSH_CMDPARMS is not set
 CONFIG_ARCH="arm"
-CONFIG_ARCH_BOARD="rddrone-uavcan146"
-CONFIG_ARCH_BOARD_RDDRONE_UAVCAN146=y
+CONFIG_ARCH_BOARD="ucans32k146"
+CONFIG_ARCH_BOARD_UCANS32K146=y
 CONFIG_ARCH_CHIP="s32k1xx"
 CONFIG_ARCH_CHIP_S32K146=y
 CONFIG_ARCH_CHIP_S32K14X=y
