+++ boards/arm/s32k1xx/ucans32k146/scripts/s32k146.cfg
