+++ boards/arm/s32k1xx/ucans32k146/src/s32k1xx_bringup.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * boards/arm/s32k1xx/rddrone-uavcan146/src/s32k1xx_bringup.c
+ * boards/arm/s32k1xx/ucans32k146/src/s32k1xx_bringup.c
  *
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
@@ -47,7 +47,7 @@
 #  include "s32k1xx_eeeprom.h"
 #endif
 
-#include "rddrone-uavcan146.h"
+#include "ucans32k146.h"
 
 /****************************************************************************
  * Public Functions
