+++ boards/arm/s32k1xx/ucans32k146/src/s32k1xx_clockconfig.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * boards/arm/s32k1xx/rddrone-uavcan146/src/s32k1xx_clockconfig.c
+ * boards/arm/s32k1xx/ucans32k146/src/s32k1xx_clockconfig.c
  *
  *   Copyright (C) 2019 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
@@ -63,7 +63,7 @@
 
 #include "s32k1xx_clockconfig.h"
 #include "s32k1xx_start.h"
-#include "rddrone-uavcan146.h"
+#include "ucans32k146.h"
 
 /****************************************************************************
  * Public Data
