+++ boards/arm/s32k1xx/ucans32k146/src/s32k1xx_periphclocks.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * boards/arm/s32k1xx/rddrone-uavcan146/src/s32k1xx_periphclocks.c
+ * boards/arm/s32k1xx/ucans32k146/src/s32k1xx_periphclocks.c
  *
  *   Copyright (C) 2019 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
@@ -59,7 +59,7 @@
 #include <nuttx/config.h>
 
 #include "s32k1xx_periphclocks.h"
-#include "rddrone-uavcan146.h"
+#include "ucans32k146.h"
 
 /****************************************************************************
  * Public Data
